+++ src/reducers/tree.js
@@ -59,11 +59,6 @@ const Tree = (state = getDefaultState(), action) => {
       return Object.assign({}, state, {
         loaded: false
       });
-    case types.CHANGE_TREE_ROOT_IDX:
-      console.warn("CHANGE_TREE_ROOT_IDX is deprecated");
-      return Object.assign({}, state, {
-        idxOfInViewRootNode: action.idxOfInViewRootNode
-      });
     case types.CHANGE_DATES_VISIBILITY_THICKNESS: /* fall-through */
     case types.UPDATE_VISIBILITY_AND_BRANCH_THICKNESS:
       const newStates = {
@@ -73,7 +68,7 @@ const Tree = (state = getDefaultState(), action) => {
         branchThicknessVersion: action.branchThicknessVersion,
         idxOfInViewRootNode: action.idxOfInViewRootNode,
         visibleStateCounts: getValuesAndCountsOfVisibleTraitsFromTree(state.nodes, action.visibility, action.stateCountAttrs)
-      }
+      };
       /* we only want to calculate totalStateCounts on the first pass */
       if (!state.loaded) {
         newStates.totalStateCounts = getAllValuesAndCountsOfTraitsFromTree(state.nodes, action.stateCountAttrs);
